Checkpoint commit
